% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboard.R
\name{billboard}
\alias{billboard}
\title{Add images to vacant facets in ggplots}
\usage{
billboard(plot, images)
}
\arguments{
\item{plot}{A ggplot object}

\item{images}{A single image grob or a list of image grobs (see examples).}
}
\description{
When there are more facets than categories of data, fill the vacant facets
with images.
}
\examples{
if (all(c(require("png"), require("jpeg")))) {
  library(ggplot2)

  img1 <- readPNG(system.file("img", "Rlogo.png", package="ggbillboard"))
  img2 <- readJPEG(system.file("img", "spongebob.png", package="ggbillboard"))
  g1 <- grid::rasterGrob(img1, interpolate=TRUE)
  g2 <- grid::rasterGrob(img2, interpolate=TRUE)

  p1 <-
    ggplot(economics_long, aes(date, value)) +
    geom_line() +
    facet_wrap(vars(variable), scales = "free_y", nrow = 2)

  billboard(p1, g1)
  billboard(p1, list(g1))

  p2 <-
    ggplot(ChickWeight, aes(Time, weight)) +
    geom_point() +
    facet_wrap(~ Diet, ncol = 3)

  billboard(p2, list(g1, g2))
  billboard(p2, list(g2, g1))
}
}
